Implementación de Spring Security y DTO de Curso
- Implementación de Spring Security para el acceso a las rutas según el Rol.
- Creación de los DTO de la entidad Curso.

+++ src/main/java/com/elearning/platform_backend/features/cursos/CursoService.java
@@ -1,43 +1,87 @@
 package com.elearning.platform_backend.features.cursos;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
+import com.elearning.platform_backend.features.usuarios.Docente;
+import com.elearning.platform_backend.features.usuarios.DocenteRepository;
+
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
 public class CursoService {
     private final CursoRepository cursoRepository;
+    private final DocenteRepository docenteRepository;
 
-    public List<Curso> listarTodos() {
-        return cursoRepository.findAll();
+
+    // Método Post para crear un nuevo Curso usando Transacional para verificar que el Docente exista
+    @Transactional
+    public CursoReaderDTO create (CursoWriterDTO dto){
+        Docente docente = docenteRepository.findById(dto.docenteID())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Docente no encontrado"));
+
+        Curso curso = CursoMapper.toEntity(dto, docente);
+        curso.setFechaCreacion(LocalDateTime.now());
+        Curso cursoGuardado = cursoRepository.save(curso);
+
+        return CursoMapper.toDto(cursoGuardado);
     }
 
-    public Curso buscarPorId(Long id) {
-        return cursoRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Curso no encontrado"));
+    // Método GET para obtener el Curso por su ID
+    public CursoReaderDTO findById(Long id){
+        Curso curso = cursoRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso no encontrado"));
+    
+        return CursoMapper.toDto(curso);
     }
 
-    public List<Curso> buscarPorTitulo(String titulo){
-        return cursoRepository.findByTitulo(titulo);
+    // Método GET para obtener a todos los Cursos
+    public List<CursoReaderDTO> findAll(){
+        if(cursoRepository.count() == 0) throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No hay cursos registrados");
+
+        return cursoRepository.findAll().stream().map(curso -> findById(curso.getId())).toList();
     }
 
-    public Curso guardar(Curso curso){
-        return cursoRepository.save(curso);
-    }
-
-    public Curso actualizar(Long id, Curso nuevo){
-        Curso curso = buscarPorId(id);
-
-        curso.setTitulo(nuevo.getTitulo());
-        curso.setDescripcion(nuevo.getDescripcion());
-
-        return cursoRepository.save(curso);
-    }
-
-    public void eliminar(Long id){
+    // Método DELETE para eliminar un Curso por ID
+    public void deleteById(Long id){
+        if(!cursoRepository.existsById(id)) throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso no existe");
         cursoRepository.deleteById(id);
+    }
+
+    // Método GET para obtener Curso por Nombre
+    public CursoReaderDTO findByTitulo(String titulo){
+        Curso curso = cursoRepository.findByTitulo(titulo).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso no encontrado"));
+
+        return CursoMapper.toDto(curso);
+    }
+
+    // Método GET para filtrado de Curso por contenido
+    public List<CursoReaderDTO> findByTituloContaining(String titulo){
+        List<Curso> cursos = cursoRepository.findByTituloContainingIgnoreCase(titulo);
+
+        if(cursos.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No se encontraron resultados");
+
+        return cursos.stream().map(CursoMapper::toDto).toList();
+    }
+
+    // Método POST para actualizar los datos del Cursos
+
+    public CursoReaderDTO update(Long id, CursoUpdateDTO dto) throws Exception {
+
+        Curso curso = cursoRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Curso no existe"));
+        Docente docente = docenteRepository.findById(dto.docenteID()).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Docente no existe"));
+        
+        curso.setTitulo(dto.titulo());
+        curso.setDescripcion(dto.descripcion());
+        curso.setDocente(docente);
+
+        return CursoMapper.toDto(cursoRepository.save(curso));
     }
 }
